refractor: TemplateCreator -> move setAlignVerticalCenter to seperate templateCreatorMutators.ts
