+++ frontend/src/util/templateCreatorUtil.ts
@@ -1,0 +1,6 @@
+export const calculateCanvasSizeFromPdfSize = (size: number) => {
+    const averageMultiplier = 0.353;
+    const canvasSize = size * averageMultiplier;
+
+    return canvasSize
+}
